fix transactions not removed from mem-pool when recieving a block

+++ org.palladiosimulator.blockchainsystems.threesim/src/main/kotlin/org/palladiosimulator/blockchainsystems/threesim/simulation/ThreesimSimulationParameters.kt
@@ -12,5 +12,5 @@ class ThreesimSimulationParameters(
   val failureThroughputThreshold: Double, // trx / s
   val shannonEntropyK: Double, // 0.0..1.0
   val nakamotoCoefficientThreshold: Double, // 0.0 .. 100.0 %
-  val reliabilityObservationTimespan: Long // ms
+  val reliabilityObservationTimespan: Double // hours
 )
